practice rounds for couples, random re-matching after practice (draw w/o replacement)
- We would also like to implement practice rounds for the couple’s part task as well. As
mentioned before, the properties of the budget lines and the nature of the task should be
identical to that of the treatment she/he was assigned to. For instance, a subject assigned to the
endogenous treatment sees in the practice rounds an allocation task and a division task.
- In the practice rounds of the couples part, the partner-subject a subject is paired with must be
different from the partner-subject of a subject in the “actual” couples part. In other words,
there is some re-matching of partners after the practice rounds of the couples part are over. For
your understanding, this is what we are telling the subjects in the instructions:
“Now the [60] rounds of Part I will start. For this purpose, you will be matched with another participant in
the room who is di¤erent from the participant you were matched with during the practice rounds of Part
I. You will stay matched with this new other participant for all [60] decision rounds of Part I.”