randomization
i) a subject needs to face the same budget lines in both parts of the
experiment (Individual and Group Parts).
ii) A group faces in each round the same budget line in the Group Part.
iii) The budget lines should differ across couples (unless they by
coincidence are the same).
iv) The order in which the lines are presented in both parts of the
experiment should differ (order should be randomized).